+++ h/crossmemory.h
@@ -144,7 +144,8 @@
 #define RC_CMS_LANC_NOT_RELEASED            93
 #define RC_CMS_STDSVC_PARM_BAD_VERSION      94
 #define RC_CMS_CONFIG_VALUE_BUF_TOO_SMALL   95
-#define RC_CMS_MAX_RC                       95
+#define RC_CMS_MODREG_FAILED                96
+#define RC_CMS_MAX_RC                       96
 
 extern const char *CMS_RC_DESCRIPTION[];
 
@@ -337,6 +338,7 @@ typedef struct CrossMemoryServer_tag {
 #define CROSS_MEMORY_SERVER_FLAG_CHECKAUTH    0x00000010
 #define CROSS_MEMORY_SERVER_FLAG_CLEAN_LPA    0x00000020
 #define CROSS_MEMORY_SERVER_FLAG_RESET_LOOKUP 0x00000040
+#define CROSS_MEMORY_SERVER_FLAG_USE_MODREG   0x00000080
   STCBase * __ptr32 base;
   CMSStarCallback * __ptr32 startCallback;
   CMSStopCallback * __ptr32 stopCallback;
@@ -460,6 +462,7 @@ ZOWE_PRAGMA_PACK_RESET
 #define CMS_SERVER_FLAG_DEV_MODE              0x00000008
 #define CMS_SERVER_FLAG_DEV_MODE_LPA          0x00000010
 #define CMS_SERVER_FLAG_RESET_LOOKUP          0x00000020
+#define CMS_SERVER_FLAG_USE_MODREG            0x00000040
 
 #define CMS_SERVICE_FLAG_NONE                 0x00000000
 #define CMS_SERVICE_FLAG_SPACE_SWITCH         0x00000001
@@ -1161,6 +1164,18 @@ typedef struct CMSDynlinkEnv_tag {
 #endif
 #define CMS_LOG_LOOKUP_ANC_RESET_WARN_MSG_TEXT  "Look-up routine anchor discard RC = %d"
 #define CMS_LOG_LOOKUP_ANC_RESET_WARN_MSG       CMS_LOG_LOOKUP_ANC_RESET_WARN_MSG_ID" "CMS_LOG_LOOKUP_ANC_RESET_WARN_MSG_TEXT
+
+#ifndef CMS_LOG_MODULE_STATUS_MSG_ID
+#define CMS_LOG_MODULE_STATUS_MSG_ID            CMS_MSG_PRFX"0258I"
+#endif
+#define CMS_LOG_MODULE_STATUS_MSG_TEXT          "Module status: %s (%p)"
+#define CMS_LOG_MODULE_STATUS_MSG               CMS_LOG_MODULE_STATUS_MSG_ID" "CMS_LOG_MODULE_STATUS_MSG_TEXT
+
+#ifndef CMS_LOG_MODREG_ADD_FAILURE_MSG_ID
+#define CMS_LOG_MODREG_ADD_FAILURE_MSG_ID       CMS_MSG_PRFX"0259E"
+#endif
+#define CMS_LOG_MODREG_ADD_FAILURE_MSG_TEXT     "Module not registered, RC = %d, RSN = 0x%016llX"
+#define CMS_LOG_MODREG_ADD_FAILURE_MSG          CMS_LOG_MODREG_ADD_FAILURE_MSG_ID" "CMS_LOG_MODREG_ADD_FAILURE_MSG_TEXT
 
 #endif /* H_CROSSMEMORY_H_ */